pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "doppelganger-ai"
version = "0.1.0"
description = "Your private, local-first AI twin — perception, memory, reasoning, voice."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["ai", "twin", "local-first", "voice", "memory", "grok"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
]

dependencies = [
    # Core
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "python-multipart>=0.0.9",

    # LLM
    "openai>=1.30.0",             # xAI uses OpenAI-compatible API

    # Memory
    "qdrant-client>=1.9.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.7.0",

    # Utils
    "psutil>=5.9.0",
    "numpy>=1.26.0",
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
voice = [
    "faster-whisper>=1.0.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "pyaudio>=0.2.14",
    "webrtcvad>=2.0.10",
]

tts-kokoro = [
    "kokoro>=0.9.0",
    "torch>=2.2.0",
]

tts-piper = [
    "piper-tts>=1.2.0",
]

csi = [
    "scapy>=2.5.0",
]

graph = [
    "neo4j>=5.19.0",
]

dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

all = [
    "doppelganger-ai[voice,tts-kokoro,graph,dev]",
]

[project.scripts]
doppelganger = "doppelganger.interfaces.cli:main"

[project.urls]
Homepage   = "https://github.com/doppelganger-ai/doppelganger"
Repository = "https://github.com/doppelganger-ai/doppelganger"
Issues     = "https://github.com/doppelganger-ai/doppelganger/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/doppelganger"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
strict = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
source = ["src"]
omit = ["tests/*"]
